Separate testing module into subpackage with individual modules. Rename TestDataTable to PlainFrame. Streamline naming conventions within tests.
