+++ tests/util/testing/test_util.py
@@ -1,0 +1,26 @@
+"""This module contains testing util tests.
+
+"""
+
+import pytest
+from pywrangler.base import BaseWrangler
+from pywrangler.util.testing.util import concretize_abstract_wrangler
+
+
+def test_concretize_abstract_wrangler():
+    class Dummy(BaseWrangler):
+        """Doc"""
+
+        @property
+        def computation_engine(self) -> str:
+            return "engine"
+
+    concrete_class = concretize_abstract_wrangler(Dummy)
+    instance = concrete_class()
+
+    assert instance.computation_engine == "engine"
+    assert instance.__doc__ == "Doc"
+    assert instance.__class__.__name__ == "Dummy"
+
+    with pytest.raises(NotImplementedError):
+        instance.preserves_sample_size
